requirements.txt
dnspython>=2.4.0
cryptography>=40.0.0
click>=8.0.0
tabulate>=0.9.0
colorama>=0.4.4
pydantic>=2.0.0
python-dateutil>=2.8.2
requests>=2.28.0
jinja2>=3.1.0
